Change: renamed robot racing files to robot fighting as I decided I wanted the minigame to instead be robot fighting.
