+++ source/core/assets/games/robot-fighting.json
