+++ source/core/src/main/com/csse3200/game/components/minigames/robotFighting/RobotFightingText.java
@@ -1,9 +1,9 @@
-package com.csse3200.game.minigames.robotRacing;
+package com.csse3200.game.components.minigames.robotFighting;
 
 import java.util.List;
 import java.util.Random;
 
-public class RobotRacingText {
+public class RobotFightingText {
     public List<String> encouragingMessages;
 
     private final Random random = new Random();
